@@ -1,18 +1,17 @@
 using System.Collections.Generic;
 
-namespace Camunda.Worker.Client
+namespace Camunda.Worker.Client;
+
+public class CompleteRequest
 {
-    public class CompleteRequest
+    public CompleteRequest(string workerId)
     {
-        public CompleteRequest(string workerId)
-        {
-            WorkerId = Guard.NotNull(workerId, nameof(workerId));
-        }
-
-        public string WorkerId { get; }
-
-        public IDictionary<string, Variable>? Variables { get; set; }
-
-        public IDictionary<string, Variable>? LocalVariables { get; set; }
+        WorkerId = Guard.NotNull(workerId, nameof(workerId));
     }
+
+    public string WorkerId { get; }
+
+    public IDictionary<string, Variable>? Variables { get; set; }
+
+    public IDictionary<string, Variable>? LocalVariables { get; set; }
 }
