@@ -1,8 +1,7 @@
-namespace Camunda.Worker.Client
+namespace Camunda.Worker.Client;
+
+public class ErrorResponse
 {
-    public class ErrorResponse
-    {
-        public string Type { get; set; } = "";
-        public string Message { get; set; } = "";
-    }
+    public string Type { get; set; } = "";
+    public string Message { get; set; } = "";
 }
