@@ -1,14 +1,13 @@
-namespace Camunda.Worker.Client
-{
-    public class ExtendLockRequest
-    {
-        public ExtendLockRequest(string workerId, int newDuration)
-        {
-            WorkerId = Guard.NotNull(workerId, nameof(workerId));
-            NewDuration = newDuration;
-        }
+namespace Camunda.Worker.Client;
 
-        public string WorkerId { get; }
-        public int NewDuration { get; }
+public class ExtendLockRequest
+{
+    public ExtendLockRequest(string workerId, int newDuration)
+    {
+        WorkerId = Guard.NotNull(workerId, nameof(workerId));
+        NewDuration = newDuration;
     }
+
+    public string WorkerId { get; }
+    public int NewDuration { get; }
 }
