@@ -2,33 +2,32 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Camunda.Worker.Client
+namespace Camunda.Worker.Client;
+
+public interface IExternalTaskClient
 {
-    public interface IExternalTaskClient
-    {
-        Task<List<ExternalTask>> FetchAndLockAsync(
-            FetchAndLockRequest request,
-            CancellationToken cancellationToken = default
-        );
+    Task<List<ExternalTask>> FetchAndLockAsync(
+        FetchAndLockRequest request,
+        CancellationToken cancellationToken = default
+    );
 
-        Task CompleteAsync(
-            string taskId, CompleteRequest request,
-            CancellationToken cancellationToken = default
-        );
+    Task CompleteAsync(
+        string taskId, CompleteRequest request,
+        CancellationToken cancellationToken = default
+    );
 
-        Task ReportFailureAsync(
-            string taskId, ReportFailureRequest request,
-            CancellationToken cancellationToken = default
-        );
+    Task ReportFailureAsync(
+        string taskId, ReportFailureRequest request,
+        CancellationToken cancellationToken = default
+    );
 
-        Task ReportBpmnErrorAsync(
-            string taskId, BpmnErrorRequest request,
-            CancellationToken cancellationToken = default
-        );
+    Task ReportBpmnErrorAsync(
+        string taskId, BpmnErrorRequest request,
+        CancellationToken cancellationToken = default
+    );
 
-        Task ExtendLockAsync(
-            string taskId, ExtendLockRequest request,
-            CancellationToken cancellationToken = default
-        );
-    }
+    Task ExtendLockAsync(
+        string taskId, ExtendLockRequest request,
+        CancellationToken cancellationToken = default
+    );
 }
