@@ -5,43 +5,42 @@ using Camunda.Worker.Client;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
-namespace Camunda.Worker
+namespace Camunda.Worker;
+
+public sealed class CompleteResult : IExecutionResult
 {
-    public sealed class CompleteResult : IExecutionResult
+    public CompleteResult()
     {
-        public CompleteResult()
+    }
+
+    public IDictionary<string, Variable>? Variables { get; set; }
+
+    public IDictionary<string, Variable>? LocalVariables { get; set; }
+
+    public async Task ExecuteResultAsync(IExternalTaskContext context)
+    {
+        var externalTask = context.Task;
+        var client = context.Client;
+
+        try
         {
+            await client.CompleteAsync(externalTask.Id, new CompleteRequest(externalTask.WorkerId)
+            {
+                Variables = Variables,
+                LocalVariables = LocalVariables,
+            });
         }
-
-        public IDictionary<string, Variable>? Variables { get; set; }
-
-        public IDictionary<string, Variable>? LocalVariables { get; set; }
-
-        public async Task ExecuteResultAsync(IExternalTaskContext context)
+        catch (ClientException e) when (e.StatusCode == HttpStatusCode.InternalServerError)
         {
-            var externalTask = context.Task;
-            var client = context.Client;
-
-            try
+            var logger = context.ServiceProvider.GetService<ILogger<CompleteResult>>();
+            logger?.LogWarning(e, "Failed completion of task {TaskId}. Reason: {Reason}",
+                externalTask.Id, e.Message
+            );
+            await client.ReportFailureAsync(externalTask.Id, new ReportFailureRequest(externalTask.WorkerId)
             {
-                await client.CompleteAsync(externalTask.Id, new CompleteRequest(externalTask.WorkerId)
-                {
-                    Variables = Variables,
-                    LocalVariables = LocalVariables,
-                });
-            }
-            catch (ClientException e) when (e.StatusCode == HttpStatusCode.InternalServerError)
-            {
-                var logger = context.ServiceProvider.GetService<ILogger<CompleteResult>>();
-                logger?.LogWarning(e, "Failed completion of task {TaskId}. Reason: {Reason}",
-                    externalTask.Id, e.Message
-                );
-                await client.ReportFailureAsync(externalTask.Id, new ReportFailureRequest(externalTask.WorkerId)
-                {
-                    ErrorMessage = e.ErrorType,
-                    ErrorDetails = e.ErrorMessage,
-                });
-            }
+                ErrorMessage = e.ErrorType,
+                ErrorDetails = e.ErrorMessage,
+            });
         }
     }
 }
