@@ -1,8 +1,7 @@
-namespace Camunda.Worker
+namespace Camunda.Worker;
+
+internal static class Constants
 {
-    internal static class Constants
-    {
-        internal const int MinimumLockDuration = 5_000;
-        internal const int MinimumParallelExecutors = 1;
-    }
+    internal const int MinimumLockDuration = 5_000;
+    internal const int MinimumParallelExecutors = 1;
 }
