@@ -1,15 +1,14 @@
-namespace Camunda.Worker.Execution
+namespace Camunda.Worker.Execution;
+
+public sealed class HandlerDescriptor
 {
-    public sealed class HandlerDescriptor
+    public HandlerDescriptor(ExternalTaskDelegate handlerDelegate, HandlerMetadata metadata)
     {
-        public HandlerDescriptor(ExternalTaskDelegate handlerDelegate, HandlerMetadata metadata)
-        {
-            HandlerDelegate = Guard.NotNull(handlerDelegate, nameof(handlerDelegate));
-            Metadata = Guard.NotNull(metadata, nameof(metadata));
-        }
-
-        public ExternalTaskDelegate HandlerDelegate { get; }
-
-        public HandlerMetadata Metadata { get; }
+        HandlerDelegate = Guard.NotNull(handlerDelegate, nameof(handlerDelegate));
+        Metadata = Guard.NotNull(metadata, nameof(metadata));
     }
+
+    public ExternalTaskDelegate HandlerDelegate { get; }
+
+    public HandlerMetadata Metadata { get; }
 }
