@@ -1,12 +1,11 @@
 using Camunda.Worker.Client;
 
-namespace Camunda.Worker.Execution
+namespace Camunda.Worker.Execution;
+
+public interface IFetchAndLockRequestProvider
 {
-    public interface IFetchAndLockRequestProvider
-    {
-        /// <summary>
-        /// This method is called in the worker before each "fetch and lock" operation
-        /// </summary>
-        FetchAndLockRequest GetRequest();
-    }
+    /// <summary>
+    /// This method is called in the worker before each "fetch and lock" operation
+    /// </summary>
+    FetchAndLockRequest GetRequest();
 }
