@@ -1,6 +1,5 @@
 using System.Threading.Tasks;
 
-namespace Camunda.Worker
-{
-    public delegate Task ExternalTaskDelegate(IExternalTaskContext context);
-}
+namespace Camunda.Worker;
+
+public delegate Task ExternalTaskDelegate(IExternalTaskContext context);
