@@ -1,41 +1,40 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
 
-namespace Camunda.Worker
+namespace Camunda.Worker;
+
+internal static class Guard
 {
-    internal static class Guard
+    [ExcludeFromCodeCoverage]
+    internal static T NotNull<T>(T parameterValue, string parameterName) where T : class
     {
-        [ExcludeFromCodeCoverage]
-        internal static T NotNull<T>(T parameterValue, string parameterName) where T : class
+        if (parameterValue == null)
         {
-            if (parameterValue == null)
-            {
-                throw new ArgumentNullException(parameterName);
-            }
-
-            return parameterValue;
+            throw new ArgumentNullException(parameterName);
         }
 
-        [ExcludeFromCodeCoverage]
-        internal static int GreaterThanOrEqual(int value, int minValue, string parameterName)
-        {
-            if (value < minValue)
-            {
-                throw new ArgumentException($"Must be greater than or equal to {minValue}", parameterName);
-            }
+        return parameterValue;
+    }
 
-            return value;
+    [ExcludeFromCodeCoverage]
+    internal static int GreaterThanOrEqual(int value, int minValue, string parameterName)
+    {
+        if (value < minValue)
+        {
+            throw new ArgumentException($"Must be greater than or equal to {minValue}", parameterName);
         }
 
-        [ExcludeFromCodeCoverage]
-        internal static string NotEmptyAndNotNull(string value, string parameterName)
-        {
-            if (string.IsNullOrEmpty(value))
-            {
-                throw new ArgumentException($"Mustn't be null or empty string", parameterName);
-            }
+        return value;
+    }
 
-            return value;
+    [ExcludeFromCodeCoverage]
+    internal static string NotEmptyAndNotNull(string value, string parameterName)
+    {
+        if (string.IsNullOrEmpty(value))
+        {
+            throw new ArgumentException($"Mustn't be null or empty string", parameterName);
         }
+
+        return value;
     }
 }
