@@ -1,9 +1,8 @@
 using System.Threading.Tasks;
 
-namespace Camunda.Worker
+namespace Camunda.Worker;
+
+public class NoneResult : IExecutionResult
 {
-    public class NoneResult : IExecutionResult
-    {
-        public Task ExecuteResultAsync(IExternalTaskContext context) => Task.CompletedTask;
-    }
+    public Task ExecuteResultAsync(IExternalTaskContext context) => Task.CompletedTask;
 }
