@@ -2,18 +2,17 @@ using System.Threading.Tasks;
 using Camunda.Worker.Execution;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Camunda.Worker.Routing
-{
-    internal static class ExternalTaskRouter
-    {
-        internal static async Task RouteAsync(IExternalTaskContext context)
-        {
-            Guard.NotNull(context, nameof(context));
-            var provider = context.ServiceProvider;
+namespace Camunda.Worker.Routing;
 
-            var endpointProvider = provider.GetRequiredService<IEndpointProvider>();
-            var externalTaskDelegate = endpointProvider.GetEndpointDelegate(context.Task);
-            await externalTaskDelegate(context);
-        }
+internal static class ExternalTaskRouter
+{
+    internal static async Task RouteAsync(IExternalTaskContext context)
+    {
+        Guard.NotNull(context, nameof(context));
+        var provider = context.ServiceProvider;
+
+        var endpointProvider = provider.GetRequiredService<IEndpointProvider>();
+        var externalTaskDelegate = endpointProvider.GetEndpointDelegate(context.Task);
+        await externalTaskDelegate(context);
     }
 }
