@@ -1,19 +1,18 @@
-namespace Camunda.Worker
+namespace Camunda.Worker;
+
+public enum VariableType
 {
-    public enum VariableType
-    {
-        Boolean,
-        Bytes,
-        Short,
-        Integer,
-        Long,
-        Double,
-        Date,
-        String,
-        Null,
-        File,
-        Object,
-        Json,
-        Xml
-    }
+    Boolean,
+    Bytes,
+    Short,
+    Integer,
+    Long,
+    Double,
+    Date,
+    String,
+    Null,
+    File,
+    Object,
+    Json,
+    Xml
 }
